introduced Data folders
moved storage classes into data folder
added pattern settings
