+++ duHast/src/duHast/Revit/Common/Objects/Data/colour_base.py
