+++ duHast/src/duHast/Revit/Common/Objects/Data/pattern_foreground.py
@@ -1,10 +1,10 @@
 """
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-A base class used to store line category overrides.
+A base class used to store pattern category overrides.
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
-Stores overrides of projection lines.
+Stores overrides of foreground patterns.
 
 """
 
@@ -25,19 +25,18 @@ Stores overrides of projection lines.
 # - Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 # - Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
 #
-# This software is provided by the copyright holder "as is" and any express or implied warranties, including, but not limited to, the implied warranties of merchantability and fitness for a particular purpose are disclaimed.
-# In no event shall the copyright holder be liable for any direct, indirect, incidental, special, exemplary, or consequential damages (including, but not limited to, procurement of substitute goods or services; loss of use, data, or profits;
+# This software is provided by the copyright holder "as is" and any express or implied warranties, including, but not limited to, the implied warranties of merchantability and fitness for a particular purpose are disclaimed. 
+# In no event shall the copyright holder be liable for any direct, indirect, incidental, special, exemplary, or consequential damages (including, but not limited to, procurement of substitute goods or services; loss of use, data, or profits; 
 # or business interruption) however caused and on any theory of liability, whether in contract, strict liability, or tort (including negligence or otherwise) arising in any way out of the use of this software, even if advised of the possibility of such damage.
 #
 #
 #
 
 
-from duHast.Revit.Common.Objects.line_graphic_base import LineGraphicBase
+from duHast.Revit.Common.Objects.Data.pattern_graphic_base import PatternGraphicBase
 
-
-class LineProjection(LineGraphicBase):
-    data_type = "line_projection"
+class PatternForeground(PatternGraphicBase):
+    data_type = "pattern_foreground"
 
     def __init__(self, j={}):
         """
@@ -48,4 +47,4 @@ class LineProjection(LineGraphicBase):
         """
 
         # store data type  in base class
-        super(LineProjection, self).__init__(data_type=self.data_type, j=j)
+        super(PatternForeground, self).__init__(data_type=self.data_type, j=j)
